Consolidate turbine models into one module
