+++ floris/simulation/__init__.py
@@ -36,7 +36,7 @@ isort:skip_file
 
 import floris.logging_manager
 
-from .base import BaseClass, BaseModel, BaseTurbineModel, State
+from .base import BaseClass, BaseModel, State
 from .turbine.turbine import (
     axial_induction,
     Ct,
