+++ floris/simulation/turbine/__init__.py
@@ -1,4 +1,18 @@
-from floris.simulation.turbine.cosine_loss_turbine import CosineLossTurbine
-from floris.simulation.turbine.simple_turbine import SimpleTurbine
+# Copyright 2021 NREL
 
-from .turbine_utilities import build_cosine_loss_turbine_dict, check_smooth_power_curve
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy of
+# the License at http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
+
+# See https://floris.readthedocs.io for documentation
+
+from floris.simulation.turbine.performance_models import (
+    CosineLossTurbine,
+    SimpleTurbine,
+)
