--- floris/simulation/turbine/simple_turbine.py
@@ -1,107 +1,0 @@
-# Copyright 2021 NREL
-
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not
-# use this file except in compliance with the License. You may obtain a copy of
-# the License at http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations under
-# the License.
-
-# See https://floris.readthedocs.io for documentation
-
-from __future__ import annotations
-
-from collections.abc import Iterable
-
-import numpy as np
-from scipy.interpolate import interp1d
-
-from floris.simulation import BaseTurbineModel
-from floris.simulation.turbine.rotor_velocity import (
-    average_velocity,
-    rotor_velocity_air_density_correction,
-)
-from floris.type_dec import (
-    NDArrayFloat,
-)
-
-
-class SimpleTurbine(BaseTurbineModel):
-    """
-    Static class defining an actuator disk turbine model that is fully aligned with the flow. No
-    handling for yaw or tilt angles.
-
-    As with all turbine submodules, implements only static power() and thrust_coefficient() methods,
-    which are called by power() and Ct() on turbine.py, respectively. This class is not intended
-    to be instantiated; it simply defines a library of static methods.
-
-    TODO: Should the turbine submodels each implement axial_induction()?
-    """
-
-    def power(
-        power_thrust_table: dict,
-        velocities: NDArrayFloat,
-        air_density: float,
-        average_method: str = "cubic-mean",
-        cubature_weights: NDArrayFloat | None = None,
-        **_ # <- Allows other models to accept other keyword arguments
-    ):
-        # Construct power interpolant
-        power_interpolator = interp1d(
-            power_thrust_table["wind_speed"],
-            power_thrust_table["power"],
-            fill_value=0.0,
-            bounds_error=False,
-        )
-
-        # Compute the power-effective wind speed across the rotor
-        rotor_average_velocities = average_velocity(
-            velocities=velocities,
-            method=average_method,
-            cubature_weights=cubature_weights,
-        )
-
-        rotor_effective_velocities = rotor_velocity_air_density_correction(
-            velocities=rotor_average_velocities,
-            air_density=air_density,
-            ref_air_density=power_thrust_table["ref_air_density"]
-        )
-
-        # Compute power
-        power = power_interpolator(rotor_effective_velocities) * 1e3 # Convert to W
-
-        return power
-
-    def thrust_coefficient(
-        power_thrust_table: dict,
-        velocities: NDArrayFloat,
-        # air_density: float,
-        average_method: str = "cubic-mean",
-        cubature_weights: NDArrayFloat | None = None,
-        **_ # <- Allows other models to accept other keyword arguments
-    ):
-        # Construct thrust coefficient interpolant
-        thrust_coefficient_interpolator = interp1d(
-            power_thrust_table["wind_speed"],
-            power_thrust_table["thrust_coefficient"],
-            fill_value=0.0001,
-            bounds_error=False,
-        )
-
-        # Compute the effective wind speed across the rotor
-        rotor_average_velocities = average_velocity(
-            velocities=velocities,
-            method=average_method,
-            cubature_weights=cubature_weights,
-        )
-
-        # TODO: Do we need an air density correction here?
-
-        thrust_coefficient = thrust_coefficient_interpolator(rotor_average_velocities)
-
-        thrust_coefficient = np.clip(thrust_coefficient, 0.0001, 0.9999)
-
-        return thrust_coefficient
